Trait syncbox::atomic::ToAtomicRepr

pub trait ToAtomicRepr: Send + Sync {
    type Repr;
    fn from_repr(raw: Self::Repr) -> Self;
    fn to_repr(self) -> Self::Repr;
}

A value that can be stored in an atomic box

Associated Types

type Repr

The representation of the value when stored in an atomic box.

Required Methods

fn from_repr(raw: Self::Repr) -> Self

Load the value from the raw representation

fn to_repr(self) -> Self::Repr

Convert the value from the raw representation