Struct syncbox::ScheduledThreadPool

pub struct ScheduledThreadPool {
    // some fields omitted
}

A thread pool that can schedule tasks to run after a given delay, or to execute periodically. Delayed tasks do not run before their associated delays, but besides that, there are no real-time guarantees about exactly when the task will run.

Methods

impl ScheduledThreadPool

fn fixed_size(size: u32) -> ScheduledThreadPool

fn single_thread() -> ScheduledThreadPool

fn schedule_ms<T: Task + 'static>(&self, delay: u32, task: T)