Struct syncbox::util::async::FutureExperimental

pub struct Future<T: Send, E: Send> {
    // some fields omitted
}

Methods

impl<T: Send, E: Send> Future<T, E>

fn pair() -> (Future<T, E>, Complete<T, E>)

fn of(val: T) -> Future<T, E>

Returns an immediately ready future containing the supplied value.

fn error(err: E) -> Future<T, E>

Returns a future that will immediately fail with the supplied error.

fn canceled() -> Future<T, E>

fn lazy<F: FnOnce() -> R + Send, R: Async<T, E>>(f: F) -> Future<T, E>

fn receive<F: FnOnce(AsyncResult<T, E>) + Send>(self, f: F)

fn await(self) -> AsyncResult<T, E>

fn map<F: FnOnce(T) -> U + Send, U: Send>(self, _f: F) -> Future<U, E>

impl<T: Send, E: Send> Future<Option<(T, Stream<T, E>)>, E>

fn as_stream(self) -> Stream<T, E>