Struct syncbox::util::async::Stream

#[must_use = "streams are lazy and do nothing unless consumed"]
pub struct Stream<T: Send, E: Send> {
    // some fields omitted
}

Methods

impl<T: Send, E: Send> Stream<T, E>

fn pair() -> (Sender<T, E>, Stream<T, E>)

Creates a new Stream, returning it with the associated Sender.

fn collect(self) -> Future<Vec<T>, E>

Asyncronously collects the items from the Stream, returning them sorted by order of arrival.

fn iter(self) -> StreamIter<T, E>

Synchronously iterate over the Stream

fn each<F: Fn(T) + Send>(self, f: F) -> Future<(), E>

fn filter<F: Fn(&T) -> bool + Send>(self, _f: F) -> Stream<T, E>

fn map<F: Fn(T) -> U + Send, U: Send>(self, f: F) -> Stream<U, E>

fn reduce<F: Fn(U, T) -> U + Send, U: Send>(self, init: U, f: F) -> Future<U, E>

fn take(self, n: u64) -> Stream<T, E>

fn take_while<F>(self, _f: F) -> Stream<T, E> where F: Fn(&T) -> bool + Send

fn take_until<A>(self, cond: A) -> Stream<T, E> where A: Async<Error=E>