Trait syncbox::util::async::AsyncExperimental

pub trait Async<T: Send, E: Send>: Send + Sized {
    fn receive<F>(self, f: F) where F: FnOnce(AsyncResult<T, E>) + Send;

    fn receive_defer<F, R>(self, f: F, run: R) where F: FnOnce(AsyncResult<T, E>) + Send, R: Run { ... }
    fn handle<F, A, U, E2>(self, cb: F) -> Future<U, E2> where F: FnOnce(AsyncResult<T, E>) -> A + Send, A: Async<U, E2>, U: Send, E2: Send { ... }
    fn and<A: Async<U, E>, U: Send>(self, next: A) -> Future<U, E> where A: Async<U, E>, U: Send { ... }
    fn and_then<F, A, U>(self, f: F) -> Future<U, E> where F: FnOnce(T) -> A + Send, A: Async<U, E>, U: Send { ... }
    fn or<A>(self, alt: A) -> Future<T, E> where A: Async<T, E> { ... }
    fn or_else<F, A>(self, f: F) -> Future<T, E> where F: FnOnce(AsyncError<E>) -> A + Send, A: Async<T, E> { ... }
}

Required Methods

fn receive<F>(self, f: F) where F: FnOnce(AsyncResult<T, E>) + Send

Invoke the callback with the resolved Async result.

Provided Methods

fn receive_defer<F, R>(self, f: F, run: R) where F: FnOnce(AsyncResult<T, E>) + Send, R: Run

Invoke the callback on the specified Run with the resolved Async result.

fn handle<F, A, U, E2>(self, cb: F) -> Future<U, E2> where F: FnOnce(AsyncResult<T, E>) -> A + Send, A: Async<U, E2>, U: Send, E2: Send

fn and<A: Async<U, E>, U: Send>(self, next: A) -> Future<U, E> where A: Async<U, E>, U: Send

If the future completes successfully, returns the complection of next.

fn and_then<F, A, U>(self, f: F) -> Future<U, E> where F: FnOnce(T) -> A + Send, A: Async<U, E>, U: Send

Also handles the Future::map case

fn or<A>(self, alt: A) -> Future<T, E> where A: Async<T, E>

fn or_else<F, A>(self, f: F) -> Future<T, E> where F: FnOnce(AsyncError<E>) -> A + Send, A: Async<T, E>