Struct syncbox::util::LinkedQueue

pub struct LinkedQueue<T> {
    // some fields omitted
}

A queue in which values are contained by a linked list.

The current implementation is based on a mutex and two condition variables. It is also mostly a placeholder until a lock-free version is implemented, so it has not been tuned for performance.

Methods

impl<T: Send> LinkedQueue<T>

fn new() -> LinkedQueue<T>

fn with_capacity(capacity: usize) -> LinkedQueue<T>

fn len(&self) -> usize

fn is_empty(&self) -> bool

fn offer(&self, e: T) -> Result<(), T>

fn put(&self, e: T)

fn poll(&self) -> Option<T>

fn take(&self) -> T

Takes from the queue, blocking until there is an element available.