Module syncbox::util::asyncExperimental

Futures & Streams

The async module contains utilities for managing asynchronous computations. These utilities are primarily based around Future and Stream types as well as functions that allow describing computations on these types.

Future

A future represents a value that will be provided sometime in the future. The value may be computed concurrently in another thread or may be provided upon completion of an asynchronous callback. The abstraction allows describing computations to perform on the value once it is realized as well as how to handle errors.

One way to think of a Future is as a Result where the value is asynchronously computed.

Stream

A stream is like a feature, except that instead of representing a single value, it represents a sequence of values.

Structs

Complete
An object that is used to fulfill or reject an associated Future.

Future

Generate

Stream

StreamIter

Enums

AsyncError

Traits

Async

Cancel

Select

ToJoin

Functions

join

select

Type Definitions

AsyncResult