Struct syncbox::util::async::Future

#[must_use = "futures are lazy and do nothing unless consumed"]
pub struct Future<T: Send, E: Send> {
    // some fields omitted
}

Methods

impl<T: Send, E: Send> Future<T, E>

fn pair() -> (Complete<T, E>, Future<T, E>)

fn of(val: T) -> Future<T, E>

Returns a future that will immediately succeed with the supplied value.

use syncbox::util::async::*;

Future::<i32, ()>::of(1).and_then(|val| {
    assert!(val == 1);
    Ok(val + 1)
});

fn error(err: E) -> Future<T, E>

Returns a future that will immediately fail with the supplied error.

use syncbox::util::async::*;
use syncbox::util::async::AsyncError::*;

Future::error("hi").or_else(|err| {
    match err {
        ExecutionError(e) => assert!(e == "hi"),
        CancellationError => unreachable!()
    }

    Ok(())
});

fn canceled() -> Future<T, E>

Returns a future that will immediately be cancelled

use syncbox::util::async::*;
use syncbox::util::async::AsyncError::*;

Future::<&'static str, ()>::canceled().or_else(|err| {
    match err {
        ExecutionError(e) => unreachable!(),
        CancellationError => assert!(true)
    }

    Ok("handled")
});

fn lazy<F, R>(f: F) -> Future<T, E> where F: FnOnce() -> R + Send, R: Async<Value=T, Error=E>

Returns a future that won't kick off its async action until a consumer registers interest.

use syncbox::util::async::*;

let post = Future::lazy(|| {
    // Imagine a call to an HTTP lib, like so:
    // http::get("/posts/1")
    Ok("HTTP response")
});

// the HTTP request has not happened yet

// later...

post.and_then(|p| {
    println!("{:?}", p);
});
// the HTTP request has now happened

fn map<F, U>(self, f: F) -> Future<U, E> where F: FnOnce(T) -> U + Send, U: Send

impl<T: Send + 'static> Future<T, ()>

fn spawn<F>(f: F) -> Future<T, ()> where F: FnOnce() -> T + Send + 'static

Returns a Future representing the completion of the given closure. The closure will be executed on a newly spawned thread.

use syncbox::util::async::*;

let future = Future::spawn(|| {
    // Represents an expensive computation
    (0..100).fold(0, |v, i| v + 1)
});

assert_eq!(100, future.await().unwrap());

impl<T: Send, E: Send> Future<Option<(T, Stream<T, E>)>, E>

fn as_stream(self) -> Stream<T, E>

An adapter that converts any future into a one-value stream