Module syncbox::util::atomic

Structs

AtomicI64

AtomicIsize
A signed integer type which can be safely shared between threads.

AtomicU64

AtomicUsize
An unsigned integer type which can be safely shared between threads.

AtomicVal

Enums

Ordering
Atomic memory orderings

Traits

Atomic
An atomic box

ToAtomicRepr
A value that can be stored in an atomic box

Functions

fence
An atomic fence.